LayoutTests/fast/dom/Geolocation/success-expected.txt

Tests Geolocation success callback using the mock service.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true
PASS position.coords.latitude is mockLatitude
PASS position.coords.longitude is mockLongitude
PASS position.coords.accuracy is mockAccuracy

TEST COMPLETE
